Refactor services structure
