+++ src/app/dialog/book-purchase-dialog/book-purchase-dialog-component.ts
@@ -2,13 +2,13 @@ import { Component, inject, OnInit } from '@angular/core';
 import { CommonModule } from '@angular/common';
 import { MAT_DIALOG_DATA, MatDialogRef, MatDialogTitle, MatDialogContent, MatDialogActions, MatDialogModule } from '@angular/material/dialog';
 import { ServiceProduct } from '../../model/service-product/service-product';
-import { BudgetService } from '../../services/budget.service';
+import { BudgetService } from '../../services/order/budget.service';
 import { MatFormField, MatLabel, MatError } from "@angular/material/form-field";
 import { MatSelect, MatSelectChange } from "@angular/material/select";
 import { MatOptionModule, MatOptionSelectionChange, provideNativeDateAdapter } from "@angular/material/core";
 import { ReactiveFormsModule, FormControl, Validators } from '@angular/forms';
 import { Budget } from '../../model/budget/budget';
-import { EventService } from '../../services/event.service';
+import { EventService } from '../../services/event/event.service';
 import { Router } from '@angular/router';
 import { MatButtonModule } from '@angular/material/button';
 import { ToastService } from '../../services/utils/toast-service';
@@ -21,7 +21,7 @@ import { MatInputModule } from '@angular/material/input';
 import { NgxMatTimepickerComponent } from "ngx-mat-timepicker";
 import { NgxMatTimepickerDirective } from 'ngx-mat-timepicker';
 import { DateRangeDto } from '../../dto/utils/date-range.dto';
-import { ServiceService } from '../../services/service.service';
+import { ServiceService } from '../../services/service-product/service.service';
 import { HttpErrorResponse } from '@angular/common/http';
 import { Service } from '../../model/service-product/service';
 import { addMilliseconds, differenceInMilliseconds, endOfDay, isBefore, max, min, parse, startOfDay } from 'date-fns';
