+++ src/app/dialog/review-dialog/review-dialog.component.ts
@@ -3,7 +3,7 @@ import { FormControl, ReactiveFormsModule } from '@angular/forms';
 import { MAT_DIALOG_DATA, MatDialogRef, MatDialogContent, MatDialogActions, MatDialogModule } from '@angular/material/dialog';
 import { ReviewService } from '../../services/review/review.service';
 import { ToastService } from '../../services/utils/toast-service';
-import { AuthService } from '../../services/auth-service.service';
+import { AuthService } from '../../services/auth/auth-service.service';
 import { MatFormField, MatFormFieldModule, MatLabel } from "@angular/material/form-field";
 import { MatButtonModule } from '@angular/material/button';
 import { ReviewDto } from '../../dto/review/review.dto';
