+++ src/app/layout/all-categories/all-categories.component.ts
@@ -3,7 +3,7 @@ import { MatTableModule } from '@angular/material/table';
 import { MatIconModule } from '@angular/material/icon';
 import { MatButtonModule } from '@angular/material/button';
 import { ServiceProductCategory } from '../../model/service-product/service-product-category';
-import { ServiceProductCategoryService } from '../../services/service-product-category.service';
+import { ServiceProductCategoryService } from '../../services/service-product/service-product-category.service';
 import { Router } from '@angular/router';
 import { MatDialog } from '@angular/material/dialog';
 import { DeleteDialogComponent } from '../../dialog/delete-dialog/delete-dialog.component';
