+++ src/app/layout/auth/login/login.component.ts
@@ -1,6 +1,6 @@
 import { Component, inject } from '@angular/core';
 import { FormBuilder, FormGroup, Validators } from '@angular/forms';
-import { AuthService } from '../../../services/auth-service.service'; // Adjust path as needed
+import { AuthService } from '../../../services/auth/auth-service.service'; // Adjust path as needed
 import { ActivatedRoute, Router } from '@angular/router';
 import { CommonModule } from '@angular/common';
 import { ReactiveFormsModule } from '@angular/forms';
