+++ src/app/layout/budget/budget.component.ts
@@ -6,12 +6,12 @@ import { MatIconModule } from '@angular/material/icon';
 import { MatButtonModule } from '@angular/material/button';
 import { MatFormFieldModule } from '@angular/material/form-field';
 import { ActivatedRoute } from '@angular/router';
-import { EventService } from '../../services/event.service';
-import { ServiceProductCategoryService } from '../../services/service-product-category.service';
+import { EventService } from '../../services/event/event.service';
+import { ServiceProductCategoryService } from '../../services/service-product/service-product-category.service';
 import { MatOption } from "@angular/material/core";
 import { MatInputModule } from '@angular/material/input';
 import { MatSelectModule } from '@angular/material/select';
-import { BudgetService } from '../../services/budget.service';
+import { BudgetService } from '../../services/order/budget.service';
 import { CreateBudgetDto } from '../../dto/budget/create-budget.dto';
 import { Budget } from '../../model/budget/budget';
 import { DeleteDialogComponent } from '../../dialog/delete-dialog/delete-dialog.component';
