+++ src/app/layout/calendar/calendar.component.ts
@@ -6,7 +6,7 @@ import dayGridPlugin from '@fullcalendar/daygrid';
 import timeGridPlugin from '@fullcalendar/timegrid';
 import interactionPlugin from '@fullcalendar/interaction';
 import { CalendarService } from '../../services/user/calendar.service';
-import { AuthService } from '../../services/auth-service.service';
+import { AuthService } from '../../services/auth/auth-service.service';
 import { Router } from '@angular/router';
 import { addHours } from 'date-fns';
 import { Booking } from '../../model/budget/booking';
