+++ src/app/layout/home/home.component.ts
@@ -9,7 +9,7 @@ import { MatInputModule } from '@angular/material/input';
 import { MatOption, MatSelect } from '@angular/material/select';
 import { MatPaginatorModule, PageEvent } from '@angular/material/paginator';
 import { MatDialog, MatDialogModule } from '@angular/material/dialog';
-import { EventService } from '../../services/event.service';
+import { EventService } from '../../services/event/event.service';
 import { ActivatedRoute, Router } from '@angular/router';
 import { PagedModel } from '../../shared/model/paged-model';
 import { EventSummaryDto } from '../../dto/event/event-summary.dto';
@@ -29,7 +29,7 @@ import { environment } from '../../../environments/environment';
 import { HomeEventFilterDialogComponent } from '../../dialog/home-event-filter-dialog/home-event-filter-dialog.component';
 import { HomeServiceProductFilterDialogComponent } from '../../dialog/home-service-product-filter-dialog/home-service-product-filter-dialog.component';
 import { EventType } from '../../model/event/event-type';
-import { EventTypeService } from '../../services/event-type.service';
+import { EventTypeService } from '../../services/event/event-type.service';
 import { ServiceProductFilteringValues } from '../../dto/service-product/service-product-filtering-values.dto';
 import { HomeEventFilterDialogParams } from '../../parameters/home-event-filter-dialog-params';
 import { HomeServiceProductFilterDialogParams } from '../../parameters/home-service-product-filter-dialog-params';
@@ -40,7 +40,7 @@ import { MatMenuModule } from '@angular/material/menu';
 import { ToastService } from '../../services/utils/toast-service';
 import { ReportDialogComponent } from '../../dialog/report-dialog/report-dialog.component';
 import { ReactiveFormsModule } from '@angular/forms';
-import { AuthService } from '../../services/auth-service.service';
+import { AuthService } from '../../services/auth/auth-service.service';
 import { UserService } from '../../services/user/user.service';
 import { HttpErrorResponse } from '@angular/common/http';
 import { ImgFallbackDirective } from '../../utils/image-fallback';
