+++ src/app/layout/new-category/new-category.component.ts
@@ -8,7 +8,7 @@ import { CommonModule } from '@angular/common';
 import { ActivatedRoute } from '@angular/router';
 import { Router } from '@angular/router';
 import { MatSnackBar } from '@angular/material/snack-bar';
-import { ServiceProductCategoryService } from '../../services/service-product-category.service';
+import { ServiceProductCategoryService } from '../../services/service-product/service-product-category.service';
 import { ServiceProductCategoryDto } from '../../dto/service-product/service-product-category.dto';
 import { ServiceProductCategory } from '../../model/service-product/service-product-category';
 import { HttpErrorResponse } from '@angular/common/http';
