+++ src/app/layout/new-event/new-event.component.ts
@@ -6,8 +6,8 @@ import { ActivatedRoute } from '@angular/router';
 import { Router } from '@angular/router';
 import { MapComponent } from '../../shared/map/map.component';
 import { EventType } from '../../model/event/event-type';
-import { EventTypeService } from '../../services/event-type.service';
-import { EventService } from '../../services/event.service';
+import { EventTypeService } from '../../services/event/event-type.service';
+import { EventService } from '../../services/event/event.service';
 import { MatSnackBar } from '@angular/material/snack-bar';
 import { MatButtonModule } from '@angular/material/button';
 import { MatTooltipModule } from '@angular/material/tooltip';
