+++ src/app/layout/new-product/new-product.component.ts
@@ -4,16 +4,16 @@ import { CommonModule } from '@angular/common';
 import { MatSelectModule } from '@angular/material/select';
 import { ActivatedRoute } from '@angular/router';
 import { Router } from '@angular/router';
-import { EventTypeService } from '../../services/event-type.service';
+import { EventTypeService } from '../../services/event/event-type.service';
 import { EventType } from '../../model/event/event-type';
-import { ServiceCategoryService } from '../../services/service-category.service';
+import { ServiceCategoryService } from '../../services/service-product/service-category.service';
 import { ServiceCategory } from '../../model/service-product/service-category';
-import { ProductService } from '../../services/product.service';
+import { ProductService } from '../../services/service-product/product.service';
 import { ReactiveFormsModule } from '@angular/forms';
 import { MatSnackBar } from '@angular/material/snack-bar';
 import { forkJoin } from 'rxjs';
 import { environment } from '../../../environments/environment';
-import { ImageService } from '../../services/image.service';
+import { ImageService } from '../../services/service-product/image.service';
 import { ToastService } from '../../services/utils/toast-service';
 import { ProductDto } from '../../dto/service-product/product.dto';
 import { HttpErrorResponse } from '@angular/common/http';
