+++ src/app/role.guard.ts
@@ -1,6 +1,6 @@
 import { inject, Injectable } from '@angular/core';
 import { CanActivate, ActivatedRouteSnapshot, Router } from '@angular/router';
-import { AuthService } from './services/auth-service.service';
+import { AuthService } from './services/auth/auth-service.service';
 
 @Injectable({
   providedIn: 'root',
