+++ src/app/services/auth/auth-service.service.ts
@@ -1,10 +1,10 @@
 import { inject, Injectable } from '@angular/core';
 import { HttpClient, HttpErrorResponse } from '@angular/common/http';
 import { BehaviorSubject, Observable, tap } from 'rxjs';
-import { environment } from '../../environments/environment';
-import { LoginResponse } from '../dto/auth/login-response';
-import { QuickLoginDto } from '../dto/auth/quick-login.dto';
-import { UserRole } from '../dto/user/user-role';
+import { environment } from '../../../environments/environment';
+import { LoginResponse } from '../../dto/auth/login-response';
+import { QuickLoginDto } from '../../dto/auth/quick-login.dto';
+import { UserRole } from '../../dto/user/user-role';
 
 @Injectable({
   providedIn: 'root'
