+++ src/app/services/service-product/service-category.service.ts
@@ -1,8 +1,8 @@
 import { HttpClient } from '@angular/common/http';
 import { inject, Injectable } from '@angular/core';
 import { Observable } from 'rxjs';
-import { environment } from '../../environments/environment';
-import { ServiceCategory } from '../model/service-product/service-category';
+import { environment } from '../../../environments/environment';
+import { ServiceCategory } from '../../model/service-product/service-category';
 
 @Injectable({
   providedIn: 'root'
