+++ src/app/services/service-product/service-product-category.service.spec.ts
@@ -1,7 +1,7 @@
 import { TestBed } from '@angular/core/testing';
 
 import { ServiceProductCategoryService } from './service-product-category.service';
-import { SharedTestingModule } from '../../testing/shared-testing.module';
+import { SharedTestingModule } from '../../../testing/shared-testing.module';
 
 describe('ServiceProductCategoryService', () => {
   let service: ServiceProductCategoryService;
