+++ src/app/services/service-product/service.service.ts
@@ -1,14 +1,14 @@
 import { inject, Injectable } from '@angular/core';
-import { Service } from '../model/service-product/service';
+import { Service } from '../../model/service-product/service';
 import {HttpClient, HttpParams} from '@angular/common/http';
 import {Observable} from 'rxjs';
-import {environment} from '../../environments/environment';
-import { PagedModel } from '../shared/model/paged-model';
-import { ServiceCardDto } from '../dto/service-product/service-card-dto.dto';
-import { CreateServiceDto } from '../dto/service-product/create-service.dto';
-import { PageParams } from '../parameters/page-params';
-import { DateRangeDto } from '../dto/utils/date-range.dto';
-import { ServiceProductSummaryDto } from '../dto/service-product/service-product-summary.dto';
+import {environment} from '../../../environments/environment';
+import { PagedModel } from '../../shared/model/paged-model';
+import { ServiceCardDto } from '../../dto/service-product/service-card-dto.dto';
+import { CreateServiceDto } from '../../dto/service-product/create-service.dto';
+import { PageParams } from '../../parameters/page-params';
+import { DateRangeDto } from '../../dto/utils/date-range.dto';
+import { ServiceProductSummaryDto } from '../../dto/service-product/service-product-summary.dto';
 @Injectable({
   providedIn: 'root'
 })
@@ -37,9 +37,9 @@ export class ServiceService {
   }
 
    getMyServiceCards(
-    page: number = 0,
+    page = 0,
     size?: number,
-    name: string = '',
+    name = '',
     categoryIds?: number[],
     available?: boolean,
     minPrice?: number,
